src/tests/test_main.py
# test_main.py
import unittest
from unittest.mock import patch

from idk_some_code.main import simulate_disaster_response, initialize_victims
from idk_some_code.grid import Grid
from idk_some_code.drone import Drone


class TestMain(unittest.TestCase):
    """Integration and functionality tests for the main orchestration logic."""

    def setUp(self) -> None:
        """Common setup for the tests."""
        self.grid_size = (10, 10)
        self.grid = Grid(*self.grid_size)

    # TODO THIS NEEDS MOCK METHODS TO WORK
    # @patch('idk_some_code.main.Grid', autospec=True)  # Mocking the Grid class
    # @patch('idk_some_code.main.Drone', autospec=True)  # Mocking the Drone class
    # def test_simulate_disaster_response(self, MockDrone, MockGrid):
    #     """Tests the disaster response simulation for expected drone and grid interactions."""
    #     mock_grid_instance = MockGrid.return_value
    #     mock_drone_instance = MockDrone.return_value
    #
    #     # Example: If `simulate_disaster_response` checks if grid has victims, set return value
    #     mock_grid_instance.get_victim_positions.return_value = [(1, 1), (2, 2)]
    #
    #     # Similarly, set up any expected return values or side effects for the drone
    #     mock_drone_instance.some_method.return_value = "expected result"
    #
    #     num_drones = 3
    #     num_victims = 5
    #     simulation_time = 10
    #
    #     simulate_disaster_response(self.grid_size, num_drones, num_victims, simulation_time)
    #
    #     # Ensure that drones are initialized and interacted with the grid
    #     self.assertTrue(MockDrone.called, "Drones were not initialized in the simulation.")
    #     self.assertTrue(MockGrid.called, "Grid was not utilized during the simulation.")

    def test_initialize_victims(self):
        """Tests whether victims are initialized correctly on the grid."""
        num_victims = 5
        initialize_victims(self.grid, num_victims)
        # We expect there to be exactly `num_victims` on the grid
        victims_positions = self.grid.get_victim_positions()
        self.assertEqual(len(victims_positions), num_victims, "Not all victims were initialized on the grid.")

    # Here, you could add more tests to simulate specific scenarios,
    # such as ensuring that drones move as expected over time, that the grid updates correctly,
    # and that the final state of the grid matches expected outcomes based on the simulation parameters.
    @patch('numpy.random.randint')
    def test_victim_placement_uniqueness(self, mock_randint):
        """Ensures victims are placed in unique locations."""
        mock_randint.side_effect = [0, 0, 1, 1, 2, 2, 3, 3, 4, 4]  # Simulate "random" placements
        num_victims = 5
        initialize_victims(self.grid, num_victims)
        victims_positions = self.grid.get_victim_positions()
        unique_positions = set(victims_positions)
        self.assertEqual(len(victims_positions), len(unique_positions),
                         "Victims were placed on top of each other. Talk about a space issue!")

    # @patch('idk_some_code.main.Drone', autospec=True)
    # def test_drones_explore_as_expected(self, MockDrone):
    #     """Tests whether drones explore the grid as expected during the simulation."""
    #     num_drones = 4
    #     num_victims = 10
    #     simulation_time = 20
    #
    #     # Configure the mock to simulate drone movement and interaction
    #     def mock_assess_and_act(self, current_time):
    #         # Move in a simple pattern or stay still to simplify testing
    #         if current_time % 2 == 0:
    #             self.position = (self.position[0] + 1, self.position[1])
    #         # Add more sophisticated movement logic here as needed
    #
    #     MockDrone.side_effect = lambda grid, start_pos: unittest.mock.Mock(
    #         spec=Drone,
    #         position=start_pos,
    #         assess_and_act=mock_assess_and_act
    #     )
    #
    #     simulate_disaster_response(self.grid_size, num_drones, num_victims, simulation_time)
    #
    #     # Check if drones have moved as expected
    #     for drone in MockDrone.mock_calls:
    #         self.assertNotEqual(drone.position, (self.grid.width // 2, self.grid.height // 2),
    #                             "Some drones didn't explore as boldly as expected.")


if __name__ == "__main__":
    unittest.main()
